画像編集画面をサーバーコンポーネント化（リファクタリング）
